+++ java/javasrc/src/main/java/com/longport/trade/OrderDetail.java
@@ -9,8 +9,8 @@ public class OrderDetail {
     private String orderId;
     private OrderStatus status;
     private String stockName;
-    private long quantity;
-    private long executedQuantity;
+    private BigDecimal quantity;
+    private BigDecimal executedQuantity;
     private BigDecimal price;
     private BigDecimal executedPrice;
     private OffsetDateTime submittedAt;
@@ -56,11 +56,11 @@ public class OrderDetail {
         return stockName;
     }
 
-    public long getQuantity() {
+    public BigDecimal getQuantity() {
         return quantity;
     }
 
-    public long getExecutedQuantity() {
+    public BigDecimal getExecutedQuantity() {
         return executedQuantity;
     }
